Commit 1 06.05.21 Change UserServiceImpl.java

+++ src/main/java/com/lekasp/spring_boot/first_spring_boot_mvc_security/service/UserServiceImpl.java
@@ -39,7 +39,7 @@ public class UserServiceImpl implements UserService {
         if (optionalUser.isPresent()) {
             return optionalUser.get();
         }
-        throw new NoResultException("No User by: " + id + " present");
+        throw new NoResultException("No User with: " + id + " present");
     }
 
     @Override
@@ -49,6 +49,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(id);
+        }
+        throw new NoResultException("No User with: " + id + " present");
     }
 }
